discord-second-brain/requirements.txt
discord.py>=2.3.0
anthropic>=0.40.0
PyGitHub>=2.1.1
chromadb>=0.4.0
sentence-transformers>=2.2.2
trafilatura>=1.6.0
youtube-transcript-api>=0.6.0
yt-dlp>=2024.1.1
faster-whisper>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
